requirements.txt
# ANALYTICA Framework - Requirements
# ===================================

# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyyaml>=6.0.1

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
asyncpg>=0.29.0
redis>=5.0.1
pymongo>=4.6.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-docx>=1.1.0

# PDF & Reports
reportlab>=4.0.0
PyPDF2>=3.0.0

# Authentication & Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0

# Task Queue
celery>=5.3.0
kombu>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Integrations - Cloud
boto3>=1.34.0

# Integrations - Payments
stripe>=7.0.0

# CLI
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Utilities
python-dateutil>=2.8.2
orjson>=3.9.0
aiofiles>=23.2.0

# Development
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0
